+++ src/pages/BorrowPage/components/BorrowTable/BorrowTable.module.less
@@ -1,0 +1,14 @@
+.headerTitleRow {
+  display: flex;
+  align-items: baseline;
+  gap: 8px;
+}
+
+.checkbox {
+  height: 18px;
+  padding-left: 18px;
+}
+
+.searchSelect {
+  max-width: 750px;
+}
