+++ src/pages/BorrowPage/components/BorrowTable/columns.tsx
@@ -4,10 +4,8 @@ import Checkbox from '@banx/components/Checkbox'
 import { HeaderCell, NftInfoCell, createSolValueJSX } from '@banx/components/TableCells'
 
 import { BorrowCell } from './BorrowCell'
-import { TableNftData } from './BorrowTable'
+import { TableNftData } from './types'
 
-// import { RepayCell, StatusCell } from './TableCells'
-// import { RepayValueCell } from './TableCells/RepayValueCell'
 import styles from './BorrowTable.module.less'
 
 interface GetTableColumnsProps {
@@ -48,30 +46,6 @@ export const getTableColumns = ({
       showSorterTooltip: false,
       sorter: true,
     },
-    // {
-    //   key: 'repayValue',
-    //   dataIndex: 'repayValue',
-    //   title: () => <HeaderCell label="Debt" />,
-    //   render: (_, loan) => <RepayValueCell loan={loan} />,
-    //   showSorterTooltip: false,
-    //   sorter: true,
-    // },
-    // {
-    //   key: 'health',
-    //   dataIndex: 'health',
-    //   title: () => <HeaderCell label="Est. health" />,
-    //   render: (_, loan) => createSolValueJSX(loan.fraktBond.amountToReturn, 1e9),
-    //   showSorterTooltip: false,
-    //   sorter: true,
-    // },
-    // {
-    //   key: 'status',
-    //   dataIndex: 'status',
-    //   title: () => <HeaderCell label="Loan status" />,
-    //   render: (_, loan) => <StatusCell loan={loan} />,
-    //   showSorterTooltip: false,
-    //   sorter: true,
-    // },
     {
       title: () => <HeaderCell label="" />,
       render: (_, nft) => <BorrowCell nft={nft} disabled={!isCartEmpty} />,
