+++ src/pages/BorrowPage/components/BorrowTable/constants.ts
@@ -1,0 +1,4 @@
+export const DEFAULT_TABLE_SORT = {
+  label: 'Borrow',
+  value: 'borrow_desc',
+}
